requirements.txt
Django
djangorestframework
celery
redis
openai
langchain
chromadb
python-docx
PyPDF2